smooth animations
i play w1 on various speed setting

move speed of 4 i find the best. On speed 3.5 the normal foot unit is moving a bit too fast.

cavalry units are keeping theri 3.7, 3.2 and 2.7 value. 
without upgrade knights are as fast as foot soldiers. And with each upgrades there is significant change, that player see, and make this upgrades valuable.

vanilla attack speeds are too fast. The combat end too fast, and player dont have a time to respond. Trying to escape from combat is pointless mostly, because this would end with receiving lots of damage without retaliation. 
thus all the attacks are slow down. but all the animation still looks fine.