bleeding animation frequence
bleeding when:

0. doing nothing animation
1. move animation
2, attacking animation
3. choppnig wood animaitn
4. dying animation (hey its a feature it looks good!)

i adjust the timings and frequence.

I think the blood fountaing for both 50% and 25% health is not helping, but i need to think about some distinguish decoration - i know already it should be altered sprites for units, but its some work, 

and i need some support how to code, to avoid legal problems - an overlay? but how to erase existing pixels?
an support for externally download files? like a second wink, "you really really should check what is in gamebanana"